backend/requirements.txt
fastapi==0.109.2
uvicorn==0.26.0
python-multipart==0.0.6
python-dotenv==1.0.0
azure-cognitiveservices-speech==1.34.0
openai==1.59.7
numpy==1.26.3
scipy==1.11.4
urllib3==2.0.7
certifi==2023.7.22
ffmpeg-python==0.2.0
cffi==1.16.0
pycparser==2.22
